requirements.txt
annotated-types==0.7.0
duckdb==1.4.4
numpy==2.4.2
pandas==3.0.0
polars==1.38.1
polars-runtime-32==1.38.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
